examples/python/simple/pyproject.toml
[project]
name = "simple"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk",
    "google-adk>=0.5.0",
    "uvicorn>=0.34.2",
    "x402-a2a",
    "x402>=0.1.4",
]

[tool.uv.sources]
x402-a2a = { path = "../../../python/x402_a2a", editable = true }

[tool.ruff]
target-version = "py313"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["server"]

[project.scripts]
server = "server.main:main"

[dependency-groups]
lint = [
    "ruff>=0.13.1",
]
